backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
google-cloud-texttospeech==2.17.2
pydub==0.25.1
imageio-ffmpeg==0.6.0
python-dotenv==1.0.1
